test successful implemented

+++ utils/format_data.py
@@ -62,7 +62,34 @@ def split_doc2(data_path, out_path):
 def delete_data(path, b_range=None, e_range=None):
     all_files = load_file_name(path)[2]
     for i in tqdm(all_files, desc="delete files"):
-        delete_file(path + i)
+        if 'ignore' not in i:
+            delete_file(path + i)
+
+
+def revers_index(path):
+    data = load_txt_data(path)
+    res = []
+    for item in data:
+        raw = item.split(',')
+        doc = raw[0]
+        try:
+            abst = raw[1]
+        except IndexError:
+            continue
+        res.append('{},{}'.format(abst, doc))
+    save_txt_file(res, path)
+
+
+def filter_data(path):
+    data = load_txt_data(path)
+    res = []
+    for item in tqdm(data, desc='Filter'):
+        raw = item.split(',')
+        doc = raw[1]
+        abst = raw[0]
+        if len(doc) >= 100:
+            res.append('{},{}'.format(abst, doc))
+    save_txt_file(res, path)
 
 
 if __name__ == '__main__':
@@ -76,4 +103,4 @@ if __name__ == '__main__':
     _op = 'C:\\Users\\VY\\Desktop\\Project/data/split_data/'
     # split_doc(_mp, _op)
     _s = '。'
-    delete_data('./data/split_data/', 100000)
+    # delete_data('./data/split_data/', 100000)
